restool: fix format string for 64 bit time_t in parse_obj_command()
32 bit systems can have a libc configured with 64 bit wide time_t.
Extend the format string for time_t so that it can take 64 bit values.

common/utils.h:62:33: error: format ‘%ld’ expects argument of type ‘long int’, but argument 5 has type ‘__time_t’ {aka ‘long long int’} [-Werror=format=]
   62 |                 fprintf(stderr, "DBG: %s:%d: " _fmt, \
      |                                 ^~~~~~~~~~~~~~
restool.c:1183:9: note: in expansion of macro ‘DEBUG_PRINTF’
 1183 |         DEBUG_PRINTF("It takes %ld.%ld seconds to run command\n",
      |         ^~~~~~~~~~~~

+++ restool.c
@@ -1180,8 +1180,8 @@ static int parse_obj_command(const char *obj_type,
 
 	clock_gettime(CLOCK_REALTIME, &end_time);
 	diff_time(&start_time, &end_time, &latency);
-	DEBUG_PRINTF("It takes %ld.%ld seconds to run command\n",
-		latency.tv_sec, latency.tv_nsec);
+	DEBUG_PRINTF("It takes %lld.%09ld seconds to run command\n",
+		     (long long)latency.tv_sec, latency.tv_nsec);
 
 	if (error < 0)
 		goto out;
